+++ src/runtime_src/core/tools/common/XBUtilities.cpp
