+++ src/runtime_src/core/tools/xbmgmt2/SubCmdVersion.cpp
@@ -19,7 +19,7 @@
 #include "SubCmdVersion.h"
 #include "common/core_system.h"
 #include "gen/version.h"
-#include "XBUtilities.h"
+#include "tools/common/XBUtilities.h"
 namespace XBU = XBUtilities;
 
 // 3rd Party Library - Include Files
@@ -30,7 +30,7 @@ namespace po = boost::program_options;
 #include <iostream>
 
 // ======= R E G I S T E R   T H E   S U B C O M M A N D ======================
-#include "SubCmd.h"
+#include "tools/common/SubCmd.h"
 static const unsigned int registerResult = 
                     register_subcommand("version", 
                                         "Reports the version of the build, OS, and drivers (if present)",
