+++ src/runtime_src/core/tools/xbutil2/CMakeLists.txt
@@ -16,10 +16,10 @@ include_directories(${CMAKE_BINARY_DIR})
 file(GLOB XBUTIL_V2_BASE_FILES
   "xbutil.cpp"
   "XBUtilMain.cpp"
-  "XBUtilities.cpp"
+  "../common/XBUtilities.cpp"
   "XBReport.cpp"
   "XBDatabase.cpp"
-  "SubCmd.cpp"
+  "../common/SubCmd.cpp"
 )
 
 if(WIN32)
