--- src/runtime_src/core/tools/xbutil2/SubCmd.h
@@ -1,41 +1,0 @@
-/**
- * Copyright (C) 2019 Xilinx, Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License"). You may
- * not use this file except in compliance with the License. A copy of the
- * License is located at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-
-#ifndef __SubCmd_h_
-#define __SubCmd_h_
-
-// Please keep eternal include file dependencies to a minimum
-#include <vector>
-#include <string>
-#include <map>
-  
-typedef int (*t_subcommand)(const std::vector<std::string> &);
-  
-typedef struct {
-  std::string sSubCmd;
-  std::string sDescription;
-  t_subcommand callBackFunction;
-  bool isHidden;
-} SubCmdEntry;
-
-typedef std::map<const std::string, SubCmdEntry> SubCmdTable;
-
-unsigned int register_subcommand(const std::string &_subCommandName, const std::string &_description, t_subcommand _pSubCommand, bool _isHidden = false);
-const SubCmdEntry *getSubCmdEntry(const std::string &_sSubCmdName);
-const SubCmdTable &getSubCmdsTable();
-
-
-#endif
