+++ src/runtime_src/core/tools/xbutil2/SubCmdTop.cpp
@@ -17,7 +17,7 @@
 // ------ I N C L U D E   F I L E S -------------------------------------------
 // Local - Include Files
 #include "SubCmdTop.h"
-#include "XBUtilities.h"
+#include "tools/common/XBUtilities.h"
 namespace XBU = XBUtilities;
 
 // 3rd Party Library - Include Files
@@ -28,7 +28,7 @@ namespace po = boost::program_options;
 #include <iostream>
 
 // ======= R E G I S T E R   T H E   S U B C O M M A N D ======================
-#include "SubCmd.h"
+#include "tools/common/SubCmd.h"
 static const unsigned int registerResult = 
                     register_subcommand("top", 
                                         "<add description>",
